add the option to specify which command should run on cleanup
