+++ src/Classes/DiskHelper.php
@@ -4,7 +4,7 @@ namespace Cybex\Protector\Classes;
 
 use Cybex\Protector\Contracts\CrypterContract;
 use Cybex\Protector\Contracts\DiskHelperContract;
-use Cybex\Protector\Enums\FlushMode;
+use Cybex\Protector\Enums\ExecutionMode;
 use Cybex\Protector\Exceptions\EmptyDumpDirectoryException;
 use Cybex\Protector\Exceptions\EmptyFileWrittenException;
 use Cybex\Protector\Exceptions\FailedReadingFromDiskException;
@@ -189,8 +189,9 @@ class DiskHelper implements DiskHelperContract
     {
         $this->deleteDumpAndMetaFiles($name, $this->getLocalDisk());
 
-        if (!FlushMode::getConfiguredMode()->shouldSchedule()) {
-            $this->flushOldLocalFiles();
+        $mode = ExecutionMode::fromConfig();
+        if (!$mode->shouldSchedule()) {
+            $mode->execute(config('protector.cleanup.local_disk.command'));
         }
     }
