+++ src/Enums/ProtectorEnv.php
@@ -7,10 +7,10 @@ enum ProtectorEnv: string
     case AUTH_TOKEN = 'PROTECTOR_CLIENT_AUTH_TOKEN';
     case BASIC_AUTH = 'PROTECTOR_CLIENT_BASIC_AUTH_CREDENTIALS';
     case CHUNK_SIZE = 'PROTECTOR_SERVER_CHUNK_SIZE';
+    case CLEANUP_LOCAL_DISK_MODE = 'CLEANUP_LOCAL_DISK_MODE';
+    case CLEANUP_LOCAL_DISK_SCHEDULE = 'CLEANUP_LOCAL_DISK_SCHEDULE';
     case DUMP_ENDPOINT_ROUTE = 'PROTECTOR_SERVER_DUMP_ENDPOINT_ROUTE';
     case DUMP_ENDPOINT_URL = 'PROTECTOR_CLIENT_DUMP_ENDPOINT_URL';
-    case FLUSH_LOCAL_MODE = 'PROTECTOR_FLUSH_LOCAL_MODE';
-    case FLUSH_LOCAL_CRON = 'PROTECTOR_FLUSH_LOCAL_CRON';
     case HTTP_TIMEOUT = 'PROTECTOR_CLIENT_HTTP_TIMEOUT';
     case MAX_PACKET_LENGTH = 'PROTECTOR_DUMP_MAX_PACKET_LENGTH';
     case METADATA = 'PROTECTOR_DUMP_METADATA';
