+++ src/ProtectorServiceProvider.php
@@ -9,21 +9,20 @@ use Cybex\Protector\Classes\SchemaState\MariaDb\MariaDbSchemaStateProxy;
 use Cybex\Protector\Classes\SchemaState\MySql\MySqlSchemaStateProxy;
 use Cybex\Protector\Classes\SchemaState\Postgres\PostgresSchemaStateProxy;
 use Cybex\Protector\Classes\SodiumCrypter;
+use Cybex\Protector\Commands\CleanupLocal;
+use Cybex\Protector\Commands\CleanupStorage;
 use Cybex\Protector\Commands\CreateKeys;
 use Cybex\Protector\Commands\CreateToken;
 use Cybex\Protector\Commands\DownloadDump;
 use Cybex\Protector\Commands\ExportDump;
-use Cybex\Protector\Commands\FlushLocal;
-use Cybex\Protector\Commands\FlushStorage;
 use Cybex\Protector\Commands\ImportDump;
 use Cybex\Protector\Contracts\CrypterContract;
 use Cybex\Protector\Contracts\DiskHelperContract;
 use Cybex\Protector\Contracts\ProtectorConfigContract;
 use Cybex\Protector\Contracts\ProtectorConfiguratorContract;
 use Cybex\Protector\Contracts\SchemaStateProxyContract;
-use Cybex\Protector\Enums\FlushMode;
+use Cybex\Protector\Enums\ExecutionMode;
 use Cybex\Protector\Exceptions\UnsupportedDatabaseException;
-use Illuminate\Console\Scheduling\Schedule;
 use Illuminate\Database\Schema\MariaDbSchemaState;
 use Illuminate\Database\Schema\MySqlSchemaState;
 use Illuminate\Database\Schema\PostgresSchemaState;
@@ -75,21 +74,23 @@ class ProtectorServiceProvider extends ServiceProvider
     protected function registerCommands(): void
     {
         $this->commands([
+            CleanupLocal::class,
+            CleanupStorage::class,
             CreateKeys::class,
             CreateToken::class,
             DownloadDump::class,
             ExportDump::class,
-            FlushLocal::class,
-            FlushStorage::class,
             ImportDump::class,
         ]);
     }
 
     protected function publishConfigs(): void
     {
-        $this->publishes([
-            __DIR__ . '/../config/protector.php' => config_path('protector.php'),
-        ], ['protector', 'protector.config']);
+        foreach (['cleanup', 'client', 'dump', 'server'] as $config) {
+            $this->publishes([
+                sprintf('%s/../config/protector/%s.php', __DIR__, $config) => config_path(sprintf('protector/%s.php', $config)),
+            ], ['protector', 'protector.config', sprintf('protector.config.%s', $config)]);
+        }
     }
 
     protected function publishMigrations(): void
@@ -105,18 +106,21 @@ class ProtectorServiceProvider extends ServiceProvider
 
     protected function scheduleTasks(): void
     {
-        if (FlushMode::getConfiguredMode()->shouldSchedule()) {
-            $this->app->booted(fn() => app(Schedule::class)
-                ->command(FlushLocal::class)
-                ->cron(FlushMode::getConfiguredCron())
-                ->runInBackground(),
-            );
+        foreach (config('protector.cleanup') as $target) {
+            $mode = ExecutionMode::from($target['mode']);
+
+            if ($mode->shouldSchedule()) {
+                $mode->execute($target['command'], $target['schedule']);
+            }
         }
     }
 
     protected function mergeProtectorConfig(): void
     {
-        $this->mergeConfigFrom(__DIR__ . '/../config/protector.php', 'protector');
+        $this->mergeConfigFrom(__DIR__ . '/../config/protector/cleanup.php', 'protector.cleanup');
+        $this->mergeConfigFrom(__DIR__ . '/../config/protector/client.php', 'protector.client');
+        $this->mergeConfigFrom(__DIR__ . '/../config/protector/dump.php', 'protector.dump');
+        $this->mergeConfigFrom(__DIR__ . '/../config/protector/server.php', 'protector.server');
     }
 
     protected function mergeDiskConfig(): void
